refactor: common, domain 모듈 환경변수 정돈 및 log 패키지 이동
