--- common/src/main/java/com/pickcar/config/WebConfig.java
@@ -1,30 +1,0 @@
-package com.pickcar.config;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.web.config.EnableSpringDataWebSupport;
-import org.springframework.data.web.config.EnableSpringDataWebSupport.PageSerializationMode;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-@EnableSpringDataWebSupport(pageSerializationMode = PageSerializationMode.VIA_DTO)
-public class WebConfig implements WebMvcConfigurer {
-
-//    @Getter
-//    @Value(value = "${custom.deploy.domain}")
-//    private String deployDomain;
-//
-//    @Override
-//    public void addCorsMappings(CorsRegistry registry) {
-//        registry.addMapping("/**")
-//                .allowedOrigins(deployDomain)
-//                .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS")
-//                .allowedHeaders("*")
-//                .allowCredentials(true);
-//    }
-}
