--- domain/src/main/resources/application-local.default
@@ -1,16 +1,0 @@
-# Local 환경에서 본인이 사용할 내용을 작성합니다
-
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/pickcar_local    # TODO: Timezone 등 옵션 추가 필요
-    username: {Local DB Username}
-    password: {Local DB Password}
-  jpa:
-    hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-        show_sql: true          # TODO: 추후 log 환경으로 변경
-        format_sql: true
-        use_sql_comments: true
