--- domain/src/main/resources/application-prod.default
@@ -1,16 +1,0 @@
-# 배포 서버의 main(prod) 환경에서 사용될 환경 변수를 작성합니다.
-
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://{IP address or Domain URI}/pickcar_prod    # TODO: Timezone 등 옵션 추가 필요
-    username: {Prod Mysql DB Username}
-    password: {Prod Mysql DB Password}
-  jpa:
-    hibernate:
-      ddl-auto: create    # 현재는 배포에서도 create로 하지만, 나중에 update => 그 뒤 validate or none(최종)
-    properties:
-      hibernate:
-        show_sql: false         # TODO : 추후 log 환경으로 변경
-        format_sql: false
-        use_sql_comments: false
