--- domain/src/main/resources/application-secret.defualt
@@ -1,7 +1,0 @@
-# secret => 커스텀 환경변수 활용
-# 연관성 있는 환경변수가 많다면, 별도의 파일로 관리 해주셔야 합니다.
-
-# example (아래는 예시일 뿐입니다.)
-custom:
-  api:
-    map-api-key: {MAP API KEY}
